PaperTrader Lab/requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
plotly>=5.22
python-dotenv>=1.0
tenacity>=8.3
httpx>=0.27
pytz>=2024.1

# Backtest engine
backtrader>=1.9.78.123

# Providers
alpaca-py>=0.14
